+++ Dockerfile
@@ -1,13 +1,9 @@
-FROM  node
+FROM  node:18
 WORKDIR /app
 COPY . .
-RUN npm install
-EXPOSE 3000
-CMD ["node", '/server']
 
-# docker build .
-# docker run <images ID>
-# docker ps
-# docker stop <container id>
-# docker run -d <images ID>
-# docker run -d -p 4000:3000 <images ID>
+RUN npm install
+COPY package.json ./
+EXPOSE 3000
+CMD ["npm", "start"]
+
