+++ models/user.js
@@ -1,8 +1,5 @@
 const { Schema, model } = require('mongoose');
 const { handleMongooseError } = require('../helpers');
-const Joi = require('joi');
-
-const emailRegexp = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
 
 const userSchema = new Schema(
   {
@@ -41,35 +38,8 @@ const userSchema = new Schema(
 );
 userSchema.post('save', handleMongooseError);
 
-const registerSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const emailShema = Joi.object({
-  email: Joi.string().pattern(emailRegexp).required(),
-});
-
-const loginSchema = Joi.object({
-  email: Joi.string().pattern(emailRegexp).required(),
-  password: Joi.string().min(6).required(),
-});
-
-const updateSubscription = Joi.object({
-  subscription: Joi.string().valid('starter', 'pro', 'business').required(),
-});
-
-const schemas = {
-  registerSchema,
-  emailShema,
-  loginSchema,
-  updateSubscription,
-};
-
 const User = model('user', userSchema);
 
 module.exports = {
-  schemas,
   User,
 };
