+++ schemas/user.js
@@ -1,0 +1,31 @@
+const Joi = require('joi');
+const emailRegexp = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
+const registerSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+});
+
+const emailShema = Joi.object({
+  email: Joi.string().pattern(emailRegexp).required(),
+});
+
+const loginSchema = Joi.object({
+  email: Joi.string().pattern(emailRegexp).required(),
+  password: Joi.string().min(6).required(),
+});
+
+const updateSubscription = Joi.object({
+  subscription: Joi.string().valid('starter', 'pro', 'business').required(),
+});
+
+const schemas = {
+  registerSchema,
+  emailShema,
+  loginSchema,
+  updateSubscription,
+};
+
+module.exports = {
+  schemas,
+};
